Server side validation and annotation added. Fixed functionality and mapping. Implemented ModelAttribute

+++ src/edu/asu/safemoney/service/EmployeeUserService.java
@@ -42,7 +42,7 @@ public interface EmployeeUserService {
 	
 	public List<RequestDTO> getTransactionList(int memberId);
 	
-	public int getCustomerId (int requestId);
+	public int getCustomerId (long requestId);
 	
 	public long getAccountNo(int memberId);
